+++ src/services/GoldfishService/index.ts
@@ -1,4 +1,5 @@
 import Environment from "@environment/index";
+import HttpService from "@services/HttpService";
 
 export default class GoldfishService {
   private static async callApi(
@@ -7,14 +8,16 @@ export default class GoldfishService {
     body?: RequestInit["body"],
     headers?: RequestInit["headers"],
   ): Promise<any> {
-    return fetch(`${Environment.GOLDFISH_URL}/${route}`, {
+    const response = await HttpService.call(
+      `${Environment.GOLDFISH_URL}/${route}`,
       method,
       body,
       headers,
-    }).then((response: Response) => {
-      if (!response.ok) throw new Error(response.statusText);
-      else return response.json();
-    });
+    );
+
+    if (!response.ok) throw new Error(response.statusText);
+
+    return response.json();
   }
 
   public static getAddresses() {